Add `--use-gpg` flag to in-toto-run to sign links
Adds the command line flag `--use-gpg` to `in-toto-run` and adapts
runlib and Metablock's signing function accordingly.

If the flag is added, then the argument passed via the `--key`
option is not interpreted as path but as GPG keyid instead.

The boolean set by the flag (`use_gpg`) is passed down to the
metadata signing function, i.e. `models.Metablock.sign`, where it
controls whether `securesystemslib` or the newly added `in_toto.gpg`
module is used for signing the Metadata.