Adapt gpg tests to handle different gpg versions
Modifies the signing/verifying with default key functions to
expect an exception if the installed version of gpg is not
fully supported.

Adds additional tests to sign/verify with the keyid of the
signing key to the signing function, which should work on not
fully supported versions as well.

Also removes trailing whitespace.