Rename type field in rule_data from unpack_rule
unpack_rule returns a dictionary of rule data that used to
contain a "type" field. This is a problem if I want to create
a function that I can pass rule_data using kwargs unpacking,
i.e. **rule_data, because "type" is a Python keyword.

This commit renames type to rule_type.