Fix #80 II (WIP)
Change of plans. Completely remove in-toto.util (no need to
import sslib into in-toto namespace and document as in-toto
functions. Just document as sslib functions instead. This has the
benefit that we don't need to explain sslib names in the docstrings
(exceptions, schemas, etc...) and that we have a module less. :)

TODO:
- factor out -P in separate PR
  - revise "with
mock.patch('securesystemslib.interface.get_password'," constructs
- create issues:
  - only supports default rsa scheme on command line
  - only supports rsa and ed25519 on command line
- Add short explanation text to api.rst to explain why we document
 sslib functions.
- Proof-read
- Write commits