Move record_artifacts_to_dict to FileResolver
TODO:
- commit message
- add FileResolver docstring
- update record_artifacts_to_dict docstring
- make note of things we should fix

------

- simplifies implementation
- uses black+isort in new resolver subpackage (see tox.ini)
  black --line-length=80 in_toto/resolver
  isort --line-length=80 --profile black --project in-toto in_toto/resolver

Caveat:
normalization:
Is what we do expected behavior? Does this make exclude patterns
 with slashes platform-dependent? Check how 'gitwildmatch' treats
 dots and slashes!

exceptions:
- Shouldn't re-raise OSError as ValueError
- Shouldn't raise FormatError but ValueError

mangle:
- lstrip: decide on what to do with left strip this differs from the original
implementation as it does not raise PrefixError, also it does not handle
uri's differently, that is uri's with schemes must include the schemes in the
lstrip_prefix list, and then the schemes would no longer be in the link
metadata. it feels like lstrip_path should not be part of record_artifacts
- Probably shouldn't name.replace("\\", "/")