tries implementing verifier case-handling in in_toto_record_stop (wip)
Idea is to:
1. require a verifier and use it if signer is passed
2. infer verifier from inferred signer if signing key or gpg keyid or gpg
default is passed
3. do additional globbing + sig extraction in the latter case

The circular dependency for default gpg key (require keyid to load
link and public key, but require signature from link to get keyid) makes
this weird.

It's easier to not change the legacy code, and only implement
(1), but leave (2) and (3) as is.

Later, when we switch invocation of runlib.in_toto* functions in cli
tools, we can use or add securesystemslib API, or add self-contained
helpers in in-toto to get signer+verifier from key files or gpg keyid or
default gpg.