Change in-toto-sign key file format
blocks on #652

----

Changes `in-toto-sign` to expect pem/pkcs8 signing keys and
pem/subjectPublicKeyInfo verification keys passed with the `--key`
argument. `--key-type` is now obsolete and removed.

Otherwise the behavior of in-toto-sign remains the same.

This is part of a series of patches to prepare for removal of legacy
securesystemslib interfaces and key file formats.

**Change details**

Unlike, in-toto-verify (#652) and in-toto-run/record (#651, #649), where
new arguments were introduced for the new formats, and deprecation
warnings were added to the old arguments, in-toto-sign is changed
directly.

This is because, the main use cases for in-toto-sign have been:
- in-toto maintainers re-signing test/demo metadata
- layout-web-tool users signing online-generated layouts

Given that the layout-web-tool is currently offline for revision
(https://github.com/in-toto/layout-web-tool/pull/70) and in-toto
maintainers should be easily able to adapt, a direct change is not
expected to disrupt anyone's operations.

IMO this can even be released as part of a minor version bump.

**Test change details**

- use new pre-generated key files instead of demo key files (but keep
  using demo metadata)
- update expected keyid link file name where necessary
- remove `--key-type` in cli invocations
- remove `--key-type` -specific tests
- remove obsolete test case, which uses 4 keys

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>